refactor to use simple mocks and dockertest for unit tests

+++ internal/domain/author/usecase/usecase.go
@@ -2,31 +2,32 @@ package usecase
 
 import (
 	"context"
+	"errors"
 
 	"github.com/gmhafiz/go8/ent/gen"
 	"github.com/gmhafiz/go8/internal/domain/author"
-	authorCache "github.com/gmhafiz/go8/internal/domain/author/repository/cache"
-	"github.com/gmhafiz/go8/internal/domain/author/repository/database"
+	"github.com/gmhafiz/go8/internal/domain/author/repository"
 )
 
 type AuthorUseCase struct {
-	repo database.Repository
+	repo repository.Author
 
-	searchRepo database.Searcher
+	searchRepo repository.Searcher
 
-	cacheLRU   authorCache.AuthorLRUService
-	cacheRedis authorCache.AuthorRedisService
+	cacheLRU   repository.AuthorLRUService
+	cacheRedis repository.AuthorRedisService
 }
 
-type UseCase interface {
-	Create(ctx context.Context, a author.CreateRequest) (*gen.Author, error)
-	List(ctx context.Context, f *author.Filter) ([]*gen.Author, int64, error)
-	Read(ctx context.Context, authorID uint64) (*gen.Author, error)
+//go:generate mirip -rm -out usecase_mock.go . Author
+type Author interface {
+	Create(ctx context.Context, a *author.CreateRequest) (*gen.Author, error)
+	List(ctx context.Context, f *author.Filter) ([]*gen.Author, int, error)
+	Read(ctx context.Context, authorID uint) (*gen.Author, error)
 	Update(ctx context.Context, author *author.Update) (*gen.Author, error)
-	Delete(ctx context.Context, authorID int64) error
+	Delete(ctx context.Context, authorID uint) error
 }
 
-func New(repo database.Repository, searcher database.Searcher, cache authorCache.AuthorLRUService, redisCache authorCache.AuthorRedisService) *AuthorUseCase {
+func New(repo repository.Author, searcher repository.Searcher, cache repository.AuthorLRUService, redisCache repository.AuthorRedisService) *AuthorUseCase {
 	return &AuthorUseCase{
 		repo:       repo,
 		searchRepo: searcher,
@@ -35,23 +36,26 @@ func New(repo database.Repository, searcher database.Searcher, cache authorCache
 	}
 }
 
-func (u *AuthorUseCase) Create(ctx context.Context, r author.CreateRequest) (*gen.Author, error) {
+func (u *AuthorUseCase) Create(ctx context.Context, r *author.CreateRequest) (*gen.Author, error) {
 	return u.repo.Create(ctx, r)
 }
 
-func (u *AuthorUseCase) List(ctx context.Context, f *author.Filter) ([]*gen.Author, int64, error) {
+func (u *AuthorUseCase) List(ctx context.Context, f *author.Filter) ([]*gen.Author, int, error) {
 	if f.Base.Search {
 		return u.searchRepo.Search(ctx, f)
 	}
-	// Use cacheRedis layer which is faster. But need to remember to invalidate
-	// cache when it is stale.
+	// Use cacheRedis layer which is faster. But depends on TTL for cache
+	// invalidation. It will call repository layer if cache key is not found.
 	return u.cacheRedis.List(ctx, f)
 
 	// Call the database layer.
 	//return u.repo.List(ctx, f)
 }
 
-func (u *AuthorUseCase) Read(ctx context.Context, authorID uint64) (*gen.Author, error) {
+func (u *AuthorUseCase) Read(ctx context.Context, authorID uint) (*gen.Author, error) {
+	if authorID == 0 {
+		return nil, errors.New("ID cannot be 0")
+	}
 	return u.repo.Read(ctx, authorID)
 }
 
@@ -60,7 +64,10 @@ func (u *AuthorUseCase) Update(ctx context.Context, author *author.Update) (*gen
 	return u.cacheRedis.Update(ctx, author)
 }
 
-func (u *AuthorUseCase) Delete(ctx context.Context, authorID int64) error {
+func (u *AuthorUseCase) Delete(ctx context.Context, authorID uint) error {
+	if authorID <= 0 {
+		return errors.New("ID cannot be 0 or less")
+	}
 	// As above
 	return u.cacheRedis.Delete(ctx, authorID)
 }
